feat: rename service-api to ssh-portal-api
This change clarifies the concept of having independent microservices
implementing individual features in Lagoon.
