+++ .github/workflows/release.yaml
@@ -68,7 +68,7 @@ jobs:
     strategy:
       matrix:
         binary:
-        - service-api
+        - ssh-portal-api
         - ssh-portal
     needs:
     - tag
