+++ .goreleaser.yml
@@ -1,7 +1,7 @@
 builds:
-- id: service-api
-  dir: cmd/service-api
-  binary: service-api
+- id: ssh-portal-api
+  dir: cmd/ssh-portal-api
+  binary: ssh-portal-api
   ldflags:
   - >
     -s -w -X main.date={{.Date}} -X "main.goVersion={{.Env.GOVERSION}}"
