+++ cmd/ssh-portal-api/version.go
@@ -14,7 +14,7 @@ type VersionCmd struct{}
 
 // Run the version command to print version information.
 func (cmd *VersionCmd) Run() error {
-	fmt.Printf("Lagoon service-api %v (%v) compiled with %v on %v\n", version,
+	fmt.Printf("Lagoon ssh-portal-api %v (%v) compiled with %v on %v\n", version,
 		shortCommit, goVersion, date)
 	return nil
 }
