--- deploy/service-api/Dockerfile
@@ -1,3 +1,0 @@
-FROM alpine:3.14
-ENTRYPOINT ["/service-api"]
-COPY service-api /
