+++ deploy/ssh-portal-api/Dockerfile
@@ -1,0 +1,3 @@
+FROM alpine:3.14
+ENTRYPOINT ["/ssh-portal-api"]
+COPY ssh-portal-api /
